* 'dev' of https://github.com/Techie006/BE: (21 commits)
  [손원철]<chore>: 사용하지 않는 클래스 삭제.
  [손원철]<fix>: 임박한 나의 재료 조회 :  responose 값에 Mark_name  추가              <fix> : 나의 storage별 조회 :  유통기한만료 및 당일 재료 respone값 수정              <update> : token이 만료될시 응답 메시지 추가
  [손원철]<update,fix>  Redis (리프레쉬토큰, 나의 냉장고 조회, 오늘 할 요리 조회) 캐싱 기능구현 및 재료 Mark_name 열 추가
  [Rzreo] <feat> redis + stomp 채팅 구현
  [Rzreo] <fix> 파일 예외처리 헨들 변경
  [choonge]<feat> 프로필 이미지 수정 및 기본 이미지로 변경 추가
  [choonge07]<feat> 비밀번호 변경(마이 페이지), 북마크한 레시피 조회(마이 페이지)추가
  메세지 컨트롤러 수정
  [Rzreo] <fix> 클래스 전체 출력시 classes로 감싸지게 수정
  [Rzreo] <feat> 클래스 생성 시 이미지받음 + 시큐리티 필요없는거 풀어놓음
  [Rzreo] <feat> 테스트 채팅
  [Rzreo] <feat> 클래스 기본api
  [Rzreo] <feat> 채팅 기본기능
  데이터 전처리
  get을 post로 변경
  cors 배포주소 추가
  [Rzreo] <fix> 통계 출력 7개로 고정
  [Rzreo] <fix> 통계 api post로 수정, 엑세스 토큰 유효시간 1일로 변경
  [손원철]<chore> 메인 repo Redis 관련 내용 삭제2 .
  [손원철]<chore> 메인 repo Redis 관련 내용 삭제 .
  ...

+++ src/main/java/com/sparta/cookbank/service/DoneRecipeService.java
@@ -8,6 +8,7 @@ import com.sparta.cookbank.domain.recipe.Recipe;
 import com.sparta.cookbank.domain.myingredients.MyIngredients;
 import com.sparta.cookbank.domain.recipe.dto.RecipeFixRequestDto;
 import com.sparta.cookbank.domain.recipe.dto.RecipeFixResponseDto;
+import com.sparta.cookbank.redis.ingredient.RedisIngredientRepo;
 import com.sparta.cookbank.repository.DoneRecipeRepository;
 import com.sparta.cookbank.repository.MemberRepository;
 import com.sparta.cookbank.repository.MyIngredientsRepository;
@@ -31,24 +32,30 @@ public class DoneRecipeService {
     private final RecipeRepository recipeRepository;
     private final MyIngredientsRepository myIngredientsRepository;
     private final DoneRecipeRepository doneRecipeRepository;
+    private final RedisIngredientRepo redisIngredientRepo;
 
 
     public void UsedIngredient(Long recipeId, DoneRecipeRequestDto requestDto) {
-        for (Long id : requestDto.getIngredients_id()) {
-            MyIngredients ingredients = myIngredientsRepository.findById(id).orElseThrow(
-                    () -> new IllegalArgumentException("해당 재료가 없습니다.")
-            );
-            myIngredientsRepository.delete(ingredients);
+        if(!requestDto.getIngredients_id().isEmpty()){
+            for (Long id : requestDto.getIngredients_id()) {
+                MyIngredients ingredients = myIngredientsRepository.findById(id).orElseThrow(
+                        () -> new IllegalArgumentException("해당 재료가 없습니다.")
+                );
+                myIngredientsRepository.delete(ingredients);
+                // 레디스 캐시 초기화.
+                redisIngredientRepo.deleteAll();
+        }}
 
-            Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId()).orElseThrow(
-                    () -> new IllegalArgumentException("유저정보가 올바르지 않습니다.")
-            );
-            Recipe recipe = recipeRepository.findById(recipeId).orElseThrow(
-                    () -> new IllegalArgumentException("해당 레시피가 존재하지 않습니다.")
-            );
-            DoneRecipe doneRecipe = new DoneRecipe(member, recipe);
-            doneRecipeRepository.save(doneRecipe);
-        }
+
+        Member member = memberRepository.findById(SecurityUtil.getCurrentMemberId()).orElseThrow(
+                () -> new IllegalArgumentException("유저정보가 올바르지 않습니다.")
+        );
+        Recipe recipe = recipeRepository.findById(recipeId).orElseThrow(
+                () -> new IllegalArgumentException("해당 레시피가 존재하지 않습니다.")
+        );
+        DoneRecipe doneRecipe = new DoneRecipe(member, recipe);
+        doneRecipeRepository.save(doneRecipe);
+
     }
 
 
@@ -111,6 +118,36 @@ public class DoneRecipeService {
             fats.add(fsum);
         }
 
+        //출력 7개로 맞추기
+        int size = days.size();
+        if(size > 7){
+            days = days.subList(size-7, size);
+            carbohydrates = carbohydrates.subList(size-7,size);
+            proteins = proteins.subList(size-7,size);
+            fats = fats.subList(size-7,size);
+        }
+        else{
+            while(size<7){
+                switch (requestDto.getFilter()) {
+                    case "month":
+                        str = str.minusMonths(1);
+                        break;
+                    case "week":
+                        str = str.minusWeeks(1);
+                        break;
+                    case "day":
+                        str = str.minusDays(1);
+                        break;
+                }
+                days.add(0,str);
+                carbohydrates.add(0,0L);
+                proteins.add(0,0L);
+                fats.add(0,0L);
+
+                size++;
+            }
+        }
+
         return NutrientsRatioResponseDto.builder()
                 .days(days)
                 .carbohydrates(carbohydrates)
@@ -167,6 +204,33 @@ public class DoneRecipeService {
             }
             calories.add(csum);
         }
+
+        //출력 7개로 맞추기
+        int size = days.size();
+        if(size > 7){
+            days = days.subList(size-7, size);
+            calories = calories.subList(size-7,size);
+        }
+        else{
+            while(size<7){
+                switch (requestDto.getFilter()) {
+                    case "month":
+                        str = str.minusMonths(1);
+                        break;
+                    case "week":
+                        str = str.minusWeeks(1);
+                        break;
+                    case "day":
+                        str = str.minusDays(1);
+                        break;
+                }
+                days.add(0,str);
+                calories.add(0,0L);
+
+                size++;
+            }
+        }
+
         return CaloriesRatioResponseDto.builder()
                 .days(days)
                 .calories(calories)
@@ -196,6 +260,9 @@ public class DoneRecipeService {
         LocalDate today = LocalDate.now();
         List<DoneRecipe> todayList = doneRecipeRepository.findAllByMember_IdAndCreatedAt(member.getId(), today);
         List<DoneRecipe> yesterdayList = doneRecipeRepository.findAllByMember_IdAndCreatedAt(member.getId(), today.minusDays(1));
+        if (todayList.isEmpty() || yesterdayList.isEmpty()) {
+            throw new IllegalArgumentException("데이터를 추가해주세요!");
+        }
 
 
         return DailyRatioResponseDto.builder()
